% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/class_haplotype_graph.R
\name{buildHaplotypeGraph}
\alias{buildHaplotypeGraph}
\title{Helper function to build HaplotypeGraph object}
\usage{
buildHaplotypeGraph(phgLocalCon)
}
\arguments{
\item{phgLocalCon}{A \code{\linkS4class{PHGLocalCon}} object.}
}
\description{
Creates a \code{\linkS4class{HaplotypeGraph}} object to be used to build and store
an \code{rJava} reference object pointing to a \code{HaplotypeGraph} object
from the PHG API.
}
