% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/class_phg_metrics.R
\name{$,PHGMetrics-method}
\alias{$,PHGMetrics-method}
\title{Access Elements of PHGMetrics Objects by Name}
\usage{
\S4method{$}{PHGMetrics}(x, name)
}
\arguments{
\item{x}{A \code{PHGMetrics} object.}

\item{name}{The name of the element to access within the \code{metadata$id} field.}
}
\value{
The value associated with the specified \code{name} in the \code{metadata$id} of the
\code{PHGMetrics} object, if it exists.
}
\description{
This method allows for accessing elements of the \code{metadata} slot in
\code{PHGMetrics} objects using the \code{$} operator, specifically filtering by the
\code{id}.
}
