% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/class_all_generics.R
\name{dbName}
\alias{dbName}
\title{Return name of DB}
\usage{
dbName(object, ...)
}
\arguments{
\item{object}{an \code{rPHG} local or server connection object}

\item{...}{Additional arguments, for use in specific methods}
}
\description{
Returns the name for a given PHG database
}
