% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/class_phg_metrics.R
\name{.DollarNames.PHGMetrics}
\alias{.DollarNames.PHGMetrics}
\title{Retrieve Names Matching a Pattern from PHGMetrics Objects}
\usage{
\method{.DollarNames}{PHGMetrics}(x, pattern = "")
}
\arguments{
\item{x}{A \code{PHGMetrics} object.}

\item{pattern}{A regular expression pattern to match against the \code{id} values.}
}
\value{
A character vector of names matching the pattern.
}
\description{
This function extracts all the names from the \code{id} field in the \code{metadata}
slot of a \code{PHGMetrics} object that match a specified pattern.
}
