% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/class_all_generics.R, R/class_phg_dataset.R
\docType{methods}
\name{findSamplesByHaplotype}
\alias{findSamplesByHaplotype}
\alias{findSamplesByHaplotype,PHGDataSet-method}
\alias{findTaxabyHaplotype,PHGDataSet-method}
\title{Find taxa by haplotype IDs}
\usage{
findSamplesByHaplotype(object, ...)

\S4method{findSamplesByHaplotype}{PHGDataSet}(object, hapIds = NULL, returnType = c("list", "tibble"))
}
\arguments{
\item{object}{an \code{rPHG2} dataset}

\item{...}{Additional arguments, for use in specific methods}

\item{hapIds}{A collection of unique haplotype IDs to query. If more than one haplotype
ID is to be queried, please use a \code{character} vector.}

\item{returnType}{How do you want to results returned? Options are:
\itemize{
\item \code{"list"} for a list of haplotype ID keys and \code{character}
vectors of taxa IDs,
\item \code{"tibble"} for a \code{tibble} dataframe object that contains
haplotype ID, taxa IDs, and the number of taxa found for each
queried haplotype ID
}}
}
\description{
Return taxa IDs that contain a provided haplotype ID in \code{rPHG2}-related
datasets.
}
