% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/class_all_generics.R, R/class_phg_con_local.R
\name{hVcfFiles}
\alias{hVcfFiles}
\alias{hVcfFiles,PHGLocalCon-method}
\title{Return hVCF files}
\usage{
hVcfFiles(object, ...)

\S4method{hVcfFiles}{PHGLocalCon}(object)
}
\arguments{
\item{object}{an \code{rPHG} local connection object}

\item{...}{Additional arguments, for use in specific methods}
}
\description{
Returns a list of hVCF files.
}
