% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/class_all_generics.R, R/class_jvm_stats.R
\name{javaVersion}
\alias{javaVersion}
\alias{classPath,JvmStats-method}
\alias{javaVersion,JvmStats-method}
\title{Return Java version}
\usage{
javaVersion(object, ...)

\S4method{classPath}{JvmStats}(object)

\S4method{javaVersion}{JvmStats}(object)
}
\arguments{
\item{object}{an \code{rPHG2 JvmStats} object}

\item{...}{Additional arguments passed to methods.}
}
\description{
This generic function retrieves the Java version associated with the
specified object.
}
