% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/class_all_generics.R, R/class_jvm_stats.R
\name{jvmMemStats}
\alias{jvmMemStats}
\alias{jvmMemStats,JvmStats-method}
\title{Return JVM memory stats}
\usage{
jvmMemStats(object, ...)

\S4method{jvmMemStats}{JvmStats}(object)
}
\arguments{
\item{object}{an \code{rPHG2 JvmStats} object}

\item{...}{Additional arguments, for use in specific methods}
}
\description{
Returns the set memory stats for initialized JVM session
}
