% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/class_jvm_stats.R
\name{jvmStats}
\alias{jvmStats}
\title{Create an instance of the JvmStats class}
\usage{
jvmStats()
}
\value{
An object of class \code{JvmStats}, containing the following
fields:
\itemize{
\item \code{nJars}: The number of JAR files in the JVM classpath.
\item \code{javaVersion}: The version of Java being used.
\item \code{phgVersion}: The version of PHG (Practical Haplotype
Graph) being used.
\item \code{maxMem}: The maximum memory (in GB) that the JVM can use.
\item \code{totMem}: The total memory (in GB) allocated to the JVM.
\item \code{freeMem}: The amount of free memory (in GB) in the JVM.
\item \code{allocMem}: The amount of allocated memory (in GB) in the
JVM.
}
}
\description{
Collects JVM statistics such as the number of JAR files loaded, Java
version, PHG version, and memory usage. Returns an instance of the
\code{JvmStats} class populated with these statistics.
}
\details{
This function requires that the JVM is initialized. If the JVM is not
initialized, it will throw an error. Use the \code{initPhg()} function
to initialize the JVM with the PHG library path before calling this
function.

The function retrieves the number of JAR files in the classpath, the
version of Java and PHG being used, and memory statistics including
maximum memory, total memory, free memory, and allocated memory (all in
gigabytes).
}
\examples{
\dontrun{
# Ensure that the JVM is initialized with PHGv2 library path
initPhg("/path/to/phgv2")

# Collect JVM statistics
stats <- jvmStats()
show(stats)
}

}
\seealso{
\code{\link{JvmStats-class}}
}
