% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/link_conda_bin.R
\name{linkCondaBin}
\alias{linkCondaBin}
\title{Link a Conda environment binary to an R option}
\usage{
linkCondaBin(condaPath, envName = "phgv2-conda", bin = "agc", verbose = TRUE)
}
\arguments{
\item{condaPath}{A character string for the Conda installation path.}

\item{envName}{A character string for the environment name. Defaults to \code{"phgv2-conda"}.}

\item{bin}{A character string for the binary to link. Defaults to \code{"agc"}.}

\item{verbose}{A logical indicating whether to print a message after setting the option.
Defaults to \code{TRUE}.}
}
\value{
The function returns nothing. It is called for its side effects, namely
setting an R option and optionally printing a message.
}
\description{
This function checks that a specified binary exists within a Conda
environment and then sets an R option pointing to it. By default,
the option name is constructed as \code{"phgv2_<bin>_path"}.
}
\details{
This function calls \code{validateBinary} to ensure the binary
exists. Then it constructs an option name (\code{"phgv2_<bin>_path"})
and sets its value to the binary's path. If \code{verbose} is
\code{TRUE}, a status message is shown.
}
\examples{
\dontrun{
linkCondaBin("~/miniconda3", "phgv2-conda", "agc")
}

}
