% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/class_all_generics.R, R/class_phg_metrics.R
\name{metricsIds}
\alias{metricsIds}
\alias{metricsIds<-}
\alias{metricsIds,PHGMetrics-method}
\alias{metricsIds<-,PHGMetrics-method}
\title{Return metrics table IDs}
\usage{
metricsIds(object, ...)

metricsIds(object, ...) <- value

\S4method{metricsIds}{PHGMetrics}(object, type = NULL)

\S4method{metricsIds}{PHGMetrics}(object) <- value
}
\arguments{
\item{object}{an \code{rPHG2} metrics object}

\item{...}{Additional arguments, for use in specific methods}

\item{value}{A named vector where name(s) are the old IDs and elements are the new IDs}

\item{type}{What collection of IDs do you want to return? Defaults to \code{NULL}.}
}
\description{
Returns a PHG metrics table by given ID
}
