% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/class_all_generics.R, R/class_phg_metrics.R
\name{metricsMetaData}
\alias{metricsMetaData}
\alias{metricsMetaData,PHGMetrics-method}
\title{Return metric metadata}
\usage{
metricsMetaData(object, ...)

\S4method{metricsMetaData}{PHGMetrics}(object)
}
\arguments{
\item{object}{an \code{rPHG2} metrics object}

\item{...}{Additional arguments, for use in specific methods}
}
\description{
Returns metadata information from \code{rPHG2} metrics objects
}
