% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/class_all_generics.R, R/class_phg_metrics.R
\name{metricsTable}
\alias{metricsTable}
\alias{metricsTable<-}
\alias{metricsTable,PHGMetrics-method}
\alias{metricsTable<-,PHGMetrics-method}
\title{Return a metrics table}
\usage{
metricsTable(object, ...)

metricsTable(object, ...) <- value

\S4method{metricsTable}{PHGMetrics}(object, name = NULL, type = NULL)

\S4method{metricsTable}{PHGMetrics}(object) <- value
}
\arguments{
\item{object}{an \code{rPHG2} metrics object}

\item{...}{Additional arguments, for use in specific methods}

\item{value}{A named list where names are the metric table IDs and values are valid
metrics tables. Currently, gVCF and AnchorWave alignment files are allowed.}

\item{name}{A metric table name}

\item{type}{What collection of IDs do you want to return? Defaults to \code{NULL}.}
}
\description{
Returns a PHG metrics table by given ID
}
