% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/class_all_generics.R,
%   R/class_haplotype_graph.R, R/class_phg_dataset.R
\name{numberOfChromosomes}
\alias{numberOfChromosomes}
\alias{numberOfChromosomes,HaplotypeGraph-method}
\alias{numberOfChromosomes,PHGDataSet-method}
\title{Return number of chromosomes}
\usage{
numberOfChromosomes(object, ...)

\S4method{numberOfChromosomes}{HaplotypeGraph}(object)

\S4method{numberOfChromosomes}{PHGDataSet}(object)
}
\arguments{
\item{object}{an \code{rPHG2} local or server connection object}

\item{...}{Additional arguments, for use in specific methods}

\item{numberOfChromosomes}{A \code{character} vector of sample IDs}
}
\description{
Returns the number of chromosomes for a given object
}
