% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/class_all_generics.R, R/class_phg_dataset.R
\name{numberOfHaplotypes}
\alias{numberOfHaplotypes}
\alias{numberOfHaplotypes,PHGDataSet-method}
\title{Return number of haplotypes}
\usage{
numberOfHaplotypes(object, ...)

\S4method{numberOfHaplotypes}{PHGDataSet}(object, byRefRange = FALSE)
}
\arguments{
\item{object}{A \code{\linkS4class{PHGDataSet}} object}

\item{...}{Additional arguments, for use in specific methods}

\item{byRefRange}{If \code{TRUE}, a \code{tibble} object will created where each row is a
count of unique haplotype IDs for every reference range. If \code{FALSE}, a
global count of all haplotype IDs will be returned from the dataset.}
}
\description{
Returns the number of haplotypes either for the whole dataset or by each
individual reference range.
}
