% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/class_all_generics.R, R/class_jvm_stats.R
\name{numberOfJars}
\alias{numberOfJars}
\alias{numberOfJars,JvmStats-method}
\title{Return number of JAR files in class path}
\usage{
numberOfJars(object, ...)

\S4method{numberOfJars}{JvmStats}(object)
}
\arguments{
\item{object}{an \code{rPHG2 JvmStats} object}

\item{...}{Additional arguments, for use in specific methods}
}
\description{
Returns the number of JAR files for a given \code{JvmStats} object
}
