% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/class_all_generics.R,
%   R/class_haplotype_graph.R, R/class_phg_dataset.R
\name{numberOfRefRanges}
\alias{numberOfRefRanges}
\alias{numberOfRefRanges,HaplotypeGraph-method}
\alias{numberOfRefRanges,PHGDataSet-method}
\title{Return number of reference ranges}
\usage{
numberOfRefRanges(object, ...)

\S4method{numberOfRefRanges}{HaplotypeGraph}(object)

\S4method{numberOfRefRanges}{PHGDataSet}(object)
}
\arguments{
\item{object}{an \code{rPHG2} local or server connection object}

\item{...}{Additional arguments, for use in specific methods}

\item{numberOfRefRanges}{A \code{character} vector of sample IDs}
}
\description{
Returns the number of reference ranges for a given object
}
