% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/class_all_generics.R,
%   R/class_haplotype_graph.R, R/class_phg_dataset.R
\name{numberOfSamples}
\alias{numberOfSamples}
\alias{numberOfSamples,HaplotypeGraph-method}
\alias{numberOfSamples,PHGDataSet-method}
\title{Return number of samples}
\usage{
numberOfSamples(object, ...)

\S4method{numberOfSamples}{HaplotypeGraph}(object)

\S4method{numberOfSamples}{PHGDataSet}(object)
}
\arguments{
\item{object}{an \code{rPHG2} local or server connection object}

\item{...}{Additional arguments, for use in specific methods}

\item{numberOfSamples}{A \code{character} vector of sample IDs}
}
\description{
Returns the number of samples for a given object
}
