% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/class_all_generics.R
\name{phgConObj}
\alias{phgConObj}
\title{Return a PHG connection object}
\usage{
phgConObj(object, ...)
}
\arguments{
\item{object}{an \code{rPHG} method object}

\item{...}{Additional arguments, for use in specific methods}
}
\description{
Returns an \code{rPHG} connection object
}
