% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/class_phg_dataset.R
\docType{class}
\name{PHGDataSet-class}
\alias{PHGDataSet-class}
\alias{show,PHGDataSet-method}
\title{A PHGDataSet Class}
\usage{
\S4method{show}{PHGDataSet}(object)
}
\arguments{
\item{object}{A \code{\linkS4class{PHGDataSet}} object}
}
\description{
A \code{PHGDatSet} class is a general class for storing data found within
an hVCF file. The main components are:

\describe{
\item{samples}{Sample IDs}
\item{reference ranges}{start/stop ranges in reference genome}
\item{haplotypes}{haplotype sequence IDs (MD5 hashes)}
\item{metadata}{additional information related to haplotype IDs}
}

Prints out information regarding properties from the \code{PHGDataSet}
class to the console
}
\section{Slots}{

\describe{
\item{\code{samples}}{A character vector of sample identifiers.}

\item{\code{refRanges}}{A \code{GRanges} object from the GenomicRanges package representing reference
genomic ranges.}

\item{\code{hapIds}}{A matrix of haplotype identifiers.}

\item{\code{hapIdMeta}}{A \code{tibble} object containing metadata associated with haplotype identifiers.}

\item{\code{hapIdMetaPos}}{A \code{tibble} object containing positional metadata for haplotype identifiers.}

\item{\code{dbUri}}{A character representation of the PHGv2 DB instance}
}}

