% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/class_phg_dataset.R
\name{PHGDataSet}
\alias{PHGDataSet}
\title{PHGDataSet object constructor}
\usage{
PHGDataSet(samples, hapIds, refRanges, hapIdMeta, hapIdMetaPos, dbUri)
}
\arguments{
\item{samples}{Sample IDs}

\item{hapIds}{Haplotype sequence IDs}

\item{refRanges}{What type of PHG connection is this?}

\item{hapIdMeta}{Metadata for haplotype IDs}

\item{hapIdMetaPos}{Positional metadata for haplotype IDs}

\item{dbUri}{PHGv2 DB path}
}
\value{
A \code{PHGDataSet} object.
}
\description{
\code{PHGDataSet} is the primary container for housing hVCF data.
}
