% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/class_phg_con_local.R
\name{PHGLocalCon}
\alias{PHGLocalCon}
\title{Helper function to construct a \code{PHGLocalCon} object}
\usage{
PHGLocalCon(hVcfFiles = NULL, dbUri = NULL)
}
\arguments{
\item{hVcfFiles}{A list of \href{https://github.com/maize-genetics/phg_v2/blob/main/docs/hvcf_specifications.md}{hVCF}
files as a \code{character} vector}

\item{dbUri}{A path to a PHG configuration file}
}
\description{
Creates a \code{\linkS4class{PHGLocalCon}} object to be used to read PHG
DB data for a given set of PHG-related methods.
}
