% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/class_all_generics.R, R/class_jvm_stats.R
\name{phgVersion}
\alias{phgVersion}
\alias{phgVersion,JvmStats-method}
\title{Return PHG version}
\usage{
phgVersion(object, ...)

\S4method{phgVersion}{JvmStats}(object, granular = FALSE)
}
\arguments{
\item{object}{an \code{rPHG2 JvmStats} object}

\item{...}{Additional arguments passed to methods.}

\item{granular}{Should a \code{list} object be returned with major/minor/build/patch values?
Defaults to \code{FALSE}}
}
\description{
This generic function retrieves the PHG version associated with the
specified object.
}
