% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/class_all_generics.R, R/class_phg_metrics.R
\name{plotDot}
\alias{plotDot}
\alias{plotDot,PHGMetrics-method}
\title{Generate dot plots}
\usage{
plotDot(object, ...)

\S4method{plotDot}{PHGMetrics}(
  object,
  metricId,
  querySeqId = NULL,
  refSeqId = NULL,
  queryLab = NULL,
  refLab = NULL,
  colorId = c("strand", "score")
)
}
\arguments{
\item{object}{an \code{rPHG2} metrics object}

\item{...}{Additional arguments, for use in specific methods}

\item{metricId}{A valid metric ID \code{character} name.}

\item{querySeqId}{Vector of sequence IDs (query).}

\item{refSeqId}{Vector of sequence IDs (reference).}

\item{queryLab}{Optional label for query axis.}

\item{refLab}{Optional label for reference axis.}

\item{colorId}{How to color plots (\code{strand} or \code{score})}
}
\description{
Generates collinearity (dot) plots between 2 samples
}
