% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/vis_plot_hapids.R
\name{plotHapIds}
\alias{plotHapIds}
\title{Plot haplotype IDs by reference range}
\usage{
plotHapIds(m, sampleId = "F1_test_G1")
}
\arguments{
\item{m}{A haplotype ID matrix}

\item{sampleId}{Sample ID to compare against ancestors}
}
\description{
Plot haplotype IDs by reference range
}
