% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/class_all_generics.R, R/class_phg_dataset.R
\name{plotHaploCounts}
\alias{plotHaploCounts}
\alias{plotHaploCounts,PHGDataSet-method}
\title{Plot haplotype counts}
\usage{
plotHaploCounts(object, ...)

\S4method{plotHaploCounts}{PHGDataSet}(object, gr = NULL)
}
\arguments{
\item{object}{a \code{\linkS4class{PHGDataSet}} object}

\item{...}{Additional arguments, for use in specific methods}

\item{gr}{A \code{GenomicRanges} object for subsetting based on chromosome ID and
start/stop positions (bp). If \code{NULL}, all haplotype counts will be
plotted similar to a Manhattan plot. Defaults to \code{NULL}.}
}
\description{
Plots the counts of unique haplotype IDs found in each reference range.
}
