% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/class_all_generics.R, R/class_phg_dataset.R
\name{plotHaploDist}
\alias{plotHaploDist}
\alias{plotHaploDist,PHGDataSet-method}
\title{Plot haplotype distribution}
\usage{
plotHaploDist(object, ...)

\S4method{plotHaploDist}{PHGDataSet}(object, drop = FALSE)
}
\arguments{
\item{object}{a \code{\linkS4class{PHGDataSet}} object}

\item{...}{Additional arguments, for use in specific methods}

\item{drop}{Do you want unused unique count bins to be plotted? Defaults to \code{TRUE}.}
}
\description{
Plots a general distribution of the number of unique haplotypes in a
\code{\linkS4class{PHGDataSet}} object.
}
