% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/class_all_generics.R, R/class_haplotype_graph.R
\name{readHapIdMetaData}
\alias{readHapIdMetaData}
\alias{readHapIdMetaData,HaplotypeGraph-method}
\title{Return haplotype ID metadata}
\usage{
readHapIdMetaData(object, ...)

\S4method{readHapIdMetaData}{HaplotypeGraph}(object)
}
\arguments{
\item{object}{an \code{rPHG} local or server connection object}

\item{...}{Additional arguments, for use in specific methods}
}
\description{
Gets haplotype ID metadata for given samples and reference ranges from a
\code{HaplotypeGraph}
}
