% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/class_all_generics.R, R/class_haplotype_graph.R
\name{readHapIds}
\alias{readHapIds}
\alias{readHapIds,HaplotypeGraph-method}
\title{Return haplotype IDs}
\usage{
readHapIds(object, ...)

\S4method{readHapIds}{HaplotypeGraph}(object)
}
\arguments{
\item{object}{an \code{rPHG} local or server connection object}

\item{...}{Additional arguments, for use in specific methods}
}
\description{
Gets haplotype ID for given samples and reference ranges from a
\code{HaplotypeGraph}
}
