% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/class_all_generics.R
\name{readRefRanges}
\alias{readRefRanges}
\title{Return reference ranges}
\usage{
readRefRanges(object, ...)
}
\arguments{
\item{object}{an \code{rPHG} local or server connection object}

\item{...}{Additional arguments, for use in specific methods}
}
\description{
Get reference range data for a given PHG method
}
