% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/class_all_generics.R
\name{readSequence}
\alias{readSequence}
\title{Return DNA sequence data}
\usage{
readSequence(object, ...)
}
\arguments{
\item{object}{an \code{rPHG2} PHG data set}

\item{...}{Additional arguments, for use in specific methods}
}
\description{
Gets DNA sequence from a \code{PHGDataSet}
}
