% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/class_phg_metrics.R
\name{seqNames<-,PHGMetrics,data.frame-method}
\alias{seqNames<-,PHGMetrics,data.frame-method}
\title{Set Sequence Names for PHGMetrics Object}
\usage{
\S4method{seqNames}{PHGMetrics,data.frame}(object) <- value
}
\arguments{
\item{object}{A \code{PHGMetrics} object.}

\item{value}{A \code{data.frame} object containing \code{old_id} and \code{new_id} columns for ID
replacement.}
}
\value{
The \code{PHGMetrics} object with updated sequence names.
}
\description{
This method replaces old IDs with new IDs in the \code{PHGMetrics} object.
It ensures that both \code{metricAlign} and \code{metricGvcf} fields are updated
with the new sequence names provided in the \code{value} data frame.
}
\details{
The method first validates that the \code{value} data frame contains the
necessary columns (\code{old_id} and \code{new_id}). Then, it replaces the old IDs
with the new IDs in both \code{metricAlign} and \code{metricGvcf} fields of the
\code{PHGMetrics} object. If a replacement ID is not found, the original ID is
retained.
}
\examples{
\dontrun{
  newIds <- data.frame(
    old_id = c("old_1", "old_2", "old_3"),
    new_id = c("new_01", "new_02", "new_03")
  )

  # Assume 'met' is a PHGMetrics object
  seqNames(met) <- newIds
}

}
