% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/class_jvm_stats.R
\name{show,JvmStats-method}
\alias{show,JvmStats-method}
\title{Show method for JvmStats class}
\usage{
\S4method{show}{JvmStats}(object)
}
\arguments{
\item{object}{An object of class \code{JvmStats}.}
}
\description{
Displays JVM statistics in a formatted manner.
}
\details{
This method provides a structured and formatted output of the JVM statistics. It displays
the number of JARs, the Java version, the PHG version, and detailed memory statistics
(maximum memory, total memory, free memory, and allocated memory) in a human-readable format.
Memory values are shown in the respective units (e.g., bytes, MB, GB) defined in the object.
}
