% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/class_all_generics.R
\name{filterRefRanges}
\alias{filterRefRanges}
\title{Filter data by reference range}
\usage{
filterRefRanges(object, ...)
}
\arguments{
\item{object}{an \code{rPHG} dataset}

\item{...}{Additional arguments, for use in specific methods}
}
\description{
Filters \code{rPHG}-related datasets by reference range coordinates or ID.
}
