% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/class_all_generics.R, R/class_phg_dataset.R
\name{filterSamples}
\alias{filterSamples}
\alias{filterSamples,PHGDataSet-method}
\title{Filter data by sample IDs}
\usage{
filterSamples(object, ...)

\S4method{filterSamples}{PHGDataSet}(object, sampleIds)
}
\arguments{
\item{object}{an \code{rPHG} dataset}

\item{...}{Additional arguments, for use in specific methods}

\item{sampleIds}{A \code{character} vector of sample IDs}
}
\description{
Filters \code{rPHG}-related datasets by sample ID.
}
