% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils_general.R
\name{\%T\%}
\alias{\%T\%}
\title{User-defined function for pre-evaluated names in vectors}
\usage{
lhs \%T\% rhs
}
\arguments{
\item{lhs}{Name for element in vector. Can be either static string or object that will
need to be evaluated.}

\item{rhs}{Element in vector}
}
\value{
A named vector
}
\description{
Generates a named element. Can be used in conjunction with \code{c()} to
create a collection of named elements
}
