% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/init_phg.R
\name{initPhg}
\alias{initPhg}
\title{Initialize JVM and add class path}
\usage{
initPhg(phgPath, check = TRUE)
}
\arguments{
\item{phgPath}{Path to PHGv2 lib folder}

\item{check}{Checks for latest PHGv2 release and correct Java version. Defaults to
\code{TRUE}.}
}
\description{
This function is needed for all PHGv2 API-related tasks (e.g., building
graph objects from hVCF files or DB connections).
}
