% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils_jvm.R
\name{initPhg}
\alias{initPhg}
\title{Initialize JVM and add class path (for R&D purposes only)}
\usage{
initPhg(phgPath, verbose = TRUE)
}
\arguments{
\item{phgPath}{path to PHGv2 lib folder}

\item{verbose}{Display all JARs added classpath? Defaults to FALSE.}
}
\description{
Initialize JVM and add class path (for R&D purposes only)
}
