% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/class_jvm_stats.R
\docType{class}
\name{JvmStats-class}
\alias{JvmStats-class}
\title{A JvmStats Class}
\value{
An object of class \code{JvmStats}.
}
\description{
An S4 class to represent statistics related to the Java Virtual Machine
(JVM).
}
\section{Slots}{

\describe{
\item{\code{nJars}}{An integer representing the number of JAR files currently loaded in the JVM.}

\item{\code{javaVersion}}{A character string representing the version of Java being used.}

\item{\code{phgVersion}}{A character string representing the version of PHG (Practical Haplotype
Graph).}

\item{\code{maxMem}}{A numeric value representing the maximum amount of memory (in bytes) that
the JVM can use.}

\item{\code{totMem}}{A numeric value representing the total amount of memory (in bytes) currently
allocated to the JVM.}

\item{\code{freeMem}}{A numeric value representing the amount of free memory (in bytes) in the
JVM.}

\item{\code{allocMem}}{A numeric value representing the amount of memory (in bytes) currently
allocated by the JVM.}

\item{\code{memUnit}}{A character string representing the unit of memory (e.g., "bytes", "MB",
"GB").}
}}

