% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/class_phg_metrics.R
\docType{class}
\name{PHGMetrics-class}
\alias{PHGMetrics-class}
\title{A PHGMetrics Class}
\description{
A \code{PHGMetrics} class is a general class for storing metrics data
generated from PHGv2 operations.
}
\section{Slots}{

\describe{
\item{\code{metricAlign}}{AnchorWave alignment metrics (\code{.anchorspro})}

\item{\code{metricGvcf}}{gVCF metrics}

\item{\code{metadata}}{A \code{data.frame} object of key-value data for files}
}}

