% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/class_phg_metrics.R
\name{PHGMetrics}
\alias{PHGMetrics}
\title{PHGMetrics object constructor}
\usage{
PHGMetrics(dir = NULL)
}
\arguments{
\item{dir}{A collection of directories}
}
\value{
A \code{PHGMetrics} object.
}
\description{
\code{PHGMetrics} is the primary container for housing PHGv2 metrics data
}
