% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/class_phg_metrics.R
\name{PHGMetrics}
\alias{PHGMetrics}
\title{PHGMetrics object constructor}
\usage{
PHGMetrics(paths = NULL, metadata = NULL)
}
\arguments{
\item{metadata}{key-value metadata for files}

\item{dir}{A \code{character} vector of file and/or directory paths}
}
\value{
A \code{PHGMetrics} object.
}
\description{
\code{PHGMetrics} is the primary container for housing PHGv2 metrics data
}
