% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/class_all_generics.R
\name{plotGvcf}
\alias{plotGvcf}
\title{Generate GVCF plots}
\usage{
plotGvcf(object, ...)
}
\arguments{
\item{object}{an \code{rPHG2} metrics object}

\item{...}{Additional arguments, for use in specific methods}
}
\description{
Generates general GVCF metrics plots for genome-wide GVCF statistics
}
