% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/class_all_generics.R, R/class_phg_metrics.R
\name{plotGvcf}
\alias{plotGvcf}
\alias{plotGvcf,PHGMetrics-method}
\title{Generate GVCF plots}
\usage{
plotGvcf(object, ...)

\S4method{plotGvcf}{PHGMetrics}(
  object,
  metricId = NULL,
  f = NULL,
  nRow = NULL,
  nCol = NULL,
  tag = "A"
)
}
\arguments{
\item{object}{A \code{PHGMetrics} object containing the gVCF data.}

\item{...}{Additional arguments, for use in specific methods}

\item{metricId}{A character vector specifying the ID of the metric to be plotted. Only a
single ID is supported.}

\item{f}{A formula object defining the plot.}

\item{nRow}{An integer specifying the number of rows in the plot layout.}

\item{nCol}{An integer specifying the number of columns in the plot layout.}

\item{tag}{What tag type do you want passed to final plot?}
}
\value{
A plot object generated from the specified gVCF data and layout.
}
\description{
Generates general GVCF metrics plots for genome-wide GVCF statistics
}
