% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/class_all_generics.R,
%   R/class_haplotype_graph.R, R/class_phg_dataset.R
\name{readHapIdPosMetaData}
\alias{readHapIdPosMetaData}
\alias{readHapIdPosMetaData,HaplotypeGraph-method}
\alias{readHapIdPosMetaData,PHGDataSet-method}
\title{Return haplotype ID positional metadata}
\usage{
readHapIdPosMetaData(object, ...)

\S4method{readHapIdPosMetaData}{HaplotypeGraph}(object)

\S4method{readHapIdPosMetaData}{PHGDataSet}(object)
}
\arguments{
\item{object}{an \code{rPHG2} local or server connection object}

\item{...}{Additional arguments, for use in specific methods}
}
\description{
Gets haplotype ID positional metadata for given samples and reference ranges
from a \code{HaplotypeGraph}
}
