% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/class_all_generics.R, R/class_haplotype_graph.R
\name{readPhgDataSet}
\alias{readPhgDataSet}
\alias{readPhgDataSet,HaplotypeGraph-method}
\title{Return all hVCF data from a JVM graph}
\usage{
readPhgDataSet(object, ...)

\S4method{readPhgDataSet}{HaplotypeGraph}(object, nThreads = 2)
}
\arguments{
\item{object}{an \code{rPHG} local or server connection object}

\item{...}{Additional arguments, for use in specific methods}

\item{nThreads}{Number of threads to use for JVM evaluation. Defaults to \code{2}.}
}
\description{
Returns all haplotype, sample, and reference range IDs and relative metadata
from a \code{\linkS4class{HaplotypeGraph}} object.
}
