% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/class_all_generics.R,
%   R/class_haplotype_graph.R, R/class_phg_con_server.R, R/class_phg_dataset.R
\name{readSamples}
\alias{readSamples}
\alias{readSamples,HaplotypeGraph-method}
\alias{readSamples,PHGServerCon-method}
\alias{readSamples,PHGDataSet-method}
\title{Return samples IDs}
\usage{
readSamples(object, ...)

\S4method{readSamples}{HaplotypeGraph}(object)

\S4method{readSamples}{PHGServerCon}(object)

\S4method{readSamples}{PHGDataSet}(object)
}
\arguments{
\item{object}{an \code{rPHG2} local or server connection object}

\item{...}{Additional arguments, for use in specific methods}
}
\description{
Gets sample ID data from a \code{HaplotypeGraph}
}
