% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/class_phg_metrics.R
\name{seqNames,PHGMetrics-method}
\alias{seqNames,PHGMetrics-method}
\title{Return all contig IDs from metrics object}
\usage{
\S4method{seqNames}{PHGMetrics}(object)
}
\arguments{
\item{object}{A \code{PHGMetrics} object}
}
\value{
A vector of unique contig IDs
}
\description{
Return all contig IDs from metrics object
}
