% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/class_phg_metrics.R
\name{seqnames,PHGMetrics-method}
\alias{seqnames,PHGMetrics-method}
\title{Return all contig IDs from metrics object}
\usage{
\S4method{seqnames}{PHGMetrics}(x)
}
\arguments{
\item{x}{A \code{PHGMetrics} object}
}
\value{
A vector of unique contig IDs
}
\description{
Return all contig IDs from metrics object
}
