% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/class_all_generics.R
\name{seqNames}
\alias{seqNames}
\alias{seqNames<-}
\title{Return sequence IDs}
\usage{
seqNames(object, ...)

seqNames(object, ...) <- value
}
\arguments{
\item{object}{an \code{rPHG2} data object}

\item{...}{Additional arguments, for use in specific methods}

\item{value}{New sequence IDs}
}
\description{
Returns sequence names from an \code{rPHG2} data object
}
